+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Manage the groups of a netbird account with the `netbird::Group` resource
 - Manage the setup keys of a netbird account with the `netbird::SetupKey` resource
+- Manage the peers that joined a netbird account with the `netbird::Peer` resource
 
 ## v0.1.0 - 2026-08-11